fixing unsigned int on 42 compiler

+++ day03/ex01/FragTrap.cpp
@@ -6,7 +6,7 @@
 /*   By: rquerino <rquerino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/16 21:36:18 by rquerino          #+#    #+#             */
-/*   Updated: 2020/01/17 17:17:02 by rquerino         ###   ########.fr       */
+/*   Updated: 2020/01/17 17:51:26 by rquerino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -117,9 +117,9 @@ void            FragTrap::vaulthunter_dot_exe(std::string const & target) {
 
 // Life gain/loss related
 void            FragTrap::takeDamage(unsigned int amount) {
-    unsigned int    realDamage = amount - this->_armorDamageReduction;
+    int    realDamage = amount - this->_armorDamageReduction;
 
-    if (this->_hitPoints - realDamage < 0) {
+    if ((int) this->_hitPoints - realDamage < 0) {
         this->_hitPoints = 0;
         std::cout << "AAAAAAARGH!! <" << this->_name << "> was hit and received <";
         std::cout << amount - this->_armorDamageReduction << "> points of damage, dying." << std::endl;
